requirements.txt
beautifulsoup4==4.6.3
certifi==2018.8.13
cobble==0.1.3
Django==2.1
django-cors-headers==2.4.0
lxml==4.2.4
mammoth==0.3.25
parsimonious==0.5
python-docx==0.8.6
pytz==2018.5
pywin32==223
wincertstore==0.2
